requirements.txt
phidata
groq
packaging
python-dotenv
yfinance
duckduckgo-search
fastapi
uvicorn